Lamp WordPress Project
Apache - The Apache web server is currently the most popular web server in the world, which makes it a great default choice for hosting a website. As a DevOps / SRE engineer your responsibility to make sure that infrastructure is up and running. You don't need to develop or know the content of the website.